Start TUI port and expose interactive app
