+++ codex-rs/tui/src/main.rs
@@ -1,3 +1,4 @@
+// C# version in codex-dotnet/CodexTui/Program.cs (in progress)
 use clap::Parser;
 use codex_common::CliConfigOverrides;
 use codex_tui::Cli;
